Checkpoint on improving improvement
This patch allows unification variables to unify with fmvs,
So that from, say   alpha ~ fmv
we get              alpha := fmv

But it iterates forever on
   F alpha ~ alpha

because we flatten to
   F alpha ~ fmv
   alpha ~ fmv
then unify to
   F fmv ~ fmv
The unflatten
   F beta ~ beta
and we are back to where we started.

Dimitrios and I have new idas about saturation; I just wanted
to commit this to a branch.